Modify AuctionFixtures to relate it to the EntityFixtures
